refactor tasks a little bit
